+++ src/m2stitch/_constrained_refinement.py
