+++ src/m2stitch/_stage_model.py
