+++ src/m2stitch/_typing_utils.py
