+++ src/m2stitch/stitching.py
@@ -7,19 +7,18 @@ import numpy as np
 import pandas as pd
 from tqdm import tqdm
 
-from .constrained_refinement import refine_translations
-from .global_optimization import compute_final_position
-from .global_optimization import compute_maximum_spanning_tree
-from .stage_model import compute_image_overlap
-from .stage_model import compute_repeatability
-from .stage_model import filter_by_repeatability
-from .stage_model import replace_invalid_translations
-from .translation_computation import interpret_translation
-from .translation_computation import multi_peak_max
-from .translation_computation import pcm
-from .typing_utils import Float
-from .typing_utils import Int
-from .typing_utils import NumArray
+from ._constrained_refinement import refine_translations
+from ._global_optimization import compute_final_position
+from ._global_optimization import compute_maximum_spanning_tree
+from ._stage_model import compute_image_overlap
+from ._stage_model import compute_repeatability
+from ._stage_model import filter_by_repeatability
+from ._stage_model import replace_invalid_translations
+from ._translation_computation import interpret_translation
+from ._translation_computation import multi_peak_max
+from ._translation_computation import pcm
+from ._typing_utils import Float
+from ._typing_utils import NumArray
 
 
 def stitch_images(
